refactor(pg): load catalogs with omni's LoadMetadata
Query span and the SQL review walk-through each had their own loader from
the metadata snapshot into an omni catalog. Both now call
catalog.LoadMetadata (bytebase/omni#414); the walk-through passes Full.
