+++ backend/plugin/schema/pg/walk_through_apply_test.go
@@ -9,58 +9,8 @@ import (
 	"github.com/stretchr/testify/require"
 
 	storepb "github.com/bytebase/bytebase/backend/generated-go/store"
-	"github.com/bytebase/bytebase/backend/plugin/schema"
 	"github.com/bytebase/bytebase/backend/store/model"
 )
-
-// TestClone_BasicIsolation verifies that Clone produces an independent catalog:
-// mutations on the clone do not affect the original.
-func TestClone_BasicIsolation(t *testing.T) {
-	original := catalog.New()
-	_, err := original.Exec(`
-		CREATE SCHEMA test_schema;
-		CREATE TABLE test_schema.t1 (id serial PRIMARY KEY, name text NOT NULL);
-		CREATE INDEX t1_name_idx ON test_schema.t1 (name);
-	`, nil)
-	require.NoError(t, err)
-
-	clone := original.Clone()
-
-	// Mutate clone
-	_, err = clone.Exec(`
-		ALTER TABLE test_schema.t1 ADD COLUMN email text;
-		CREATE TABLE test_schema.t2 (id int PRIMARY KEY);
-		DROP INDEX test_schema.t1_name_idx;
-	`, nil)
-	require.NoError(t, err)
-
-	// Verify original is untouched
-	origRel := original.GetRelation("test_schema", "t1")
-	require.NotNil(t, origRel, "original t1 should still exist")
-	require.Equal(t, 2, len(origRel.Columns), "original t1 should have 2 columns (id, name), not 3")
-
-	require.Nil(t, original.GetRelation("test_schema", "t2"), "original should NOT have t2")
-
-	origIndexes := original.IndexesOf(origRel.OID)
-	found := false
-	for _, idx := range origIndexes {
-		if idx.Name == "t1_name_idx" {
-			found = true
-		}
-	}
-	require.True(t, found, "original should still have t1_name_idx")
-
-	// Verify clone has the changes
-	cloneRel := clone.GetRelation("test_schema", "t1")
-	require.NotNil(t, cloneRel)
-	require.Equal(t, 3, len(cloneRel.Columns), "clone t1 should have 3 columns")
-	require.NotNil(t, clone.GetRelation("test_schema", "t2"), "clone should have t2")
-
-	// Diff should show changes
-	diff := catalog.Diff(original, clone)
-	require.False(t, diff.IsEmpty(), "diff should not be empty")
-	require.Greater(t, len(diff.Relations), 0, "should have relation diffs")
-}
 
 // TestClone_WalkThroughIntegration tests the full walk-through flow using Clone:
 // load catalog → clone → exec user DDL on clone → diff → apply.
@@ -87,7 +37,7 @@ func TestClone_WalkThroughIntegration(t *testing.T) {
 	}
 
 	catBefore := catalog.New()
-	err := loadWalkThroughCatalog(context.Background(), catBefore, meta)
+	_, err := catBefore.LoadMetadata(context.Background(), meta, catalog.LoadMetadataOptions{Full: true})
 	require.NoError(t, err)
 
 	catAfter := catBefore.Clone()
@@ -130,89 +80,6 @@ func TestClone_WalkThroughIntegration(t *testing.T) {
 	origUsers := origMeta.GetSchemaMetadata("public").GetTable("users")
 	require.Equal(t, 2, len(origUsers.GetProto().Columns), "original should still have 2 columns")
 	require.Nil(t, origMeta.GetSchemaMetadata("public").GetTable("posts"), "original should not have posts")
-}
-
-// TestClone_SearchPath tests Clone preserves search path and session user.
-func TestClone_SearchPath(t *testing.T) {
-	meta := &metadatapb.DatabaseSchemaMetadata{
-		Name: "postgres",
-		Schemas: []*metadatapb.SchemaMetadata{
-			{Name: "public"},
-			{Name: "alice"},
-		},
-	}
-
-	catBefore := catalog.New()
-	catBefore.SetSessionUser("alice")
-	catBefore.SetSearchPath([]string{"$user", "public"})
-	err := loadWalkThroughCatalog(context.Background(), catBefore, meta)
-	require.NoError(t, err)
-
-	catAfter := catBefore.Clone()
-
-	// CREATE TABLE without schema should use search path ($user → alice)
-	_, err = catAfter.Exec(`CREATE TABLE my_table (id int);`, nil)
-	require.NoError(t, err)
-
-	// Table should be in alice schema on clone
-	require.NotNil(t, catAfter.GetRelation("alice", "my_table"))
-
-	// Original should NOT have it
-	require.Nil(t, catBefore.GetRelation("alice", "my_table"))
-
-	// Diff should show the new table
-	diff := catalog.Diff(catBefore, catAfter)
-	require.False(t, diff.IsEmpty())
-
-	foundAliceTable := false
-	for _, rel := range diff.Relations {
-		if rel.SchemaName == "alice" && rel.Name == "my_table" && rel.Action == catalog.DiffAdd {
-			foundAliceTable = true
-		}
-	}
-	require.True(t, foundAliceTable, "diff should show alice.my_table as added")
-}
-
-// TestClone_WalkThroughFunction tests the actual WalkThroughWithContext function
-// using Clone (if enabled) produces correct FinalMetadata.
-func TestClone_WalkThroughFunction(t *testing.T) {
-	meta := &metadatapb.DatabaseSchemaMetadata{
-		Name: "postgres",
-		Schemas: []*metadatapb.SchemaMetadata{
-			{
-				Name: "public",
-				Tables: []*metadatapb.TableMetadata{
-					{
-						Name: "test",
-						Columns: []*metadatapb.ColumnMetadata{
-							{Name: "id", Type: "integer", Position: 1},
-							{Name: "name", Type: "text", Position: 2, Nullable: true},
-						},
-						Indexes: []*metadatapb.IndexMetadata{
-							{Name: "test_pkey", Expressions: []string{"id"}, Unique: true, Primary: true},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	state := model.NewDatabaseMetadata(meta, nil, nil, storepb.Engine_POSTGRES, true)
-	ctx := schema.WalkThroughContext{
-		RawSQL: `CREATE TABLE public.new_table (id int PRIMARY KEY, val text);`,
-	}
-
-	advice := WalkThroughWithContext(ctx, state, nil)
-	require.Nil(t, advice, "walk-through should succeed")
-
-	// Check FinalMetadata has the new table
-	newTbl := state.GetSchemaMetadata("public").GetTable("new_table")
-	require.NotNil(t, newTbl, "new_table should exist in FinalMetadata")
-
-	// Check original table is preserved
-	origTbl := state.GetSchemaMetadata("public").GetTable("test")
-	require.NotNil(t, origTbl, "test table should still exist")
-	require.Equal(t, 2, len(origTbl.GetProto().Columns))
 }
 
 // compositeWalkThroughMetadata has an adversarial name pair (aa_nested sorts
@@ -271,21 +138,6 @@ func compositeWalkThroughMetadata() *metadatapb.DatabaseSchemaMetadata {
 	}
 }
 
-func TestWalkThroughLoadsCompositeTypes(t *testing.T) {
-	meta := compositeWalkThroughMetadata()
-
-	cat := catalog.New()
-	err := loadWalkThroughCatalog(context.Background(), cat, meta)
-	require.NoError(t, err)
-
-	require.NotNil(t, cat.GetRelation("public", "zz_base"), "composite type must be installed")
-	require.NotNil(t, cat.GetRelation("public", "aa_nested"), "nested composite type must be installed")
-	arrayOnly := cat.GetRelation("public", "aa_array_only")
-	require.NotNil(t, arrayOnly, "array-only referencing composite must be installed")
-	require.Len(t, arrayOnly.Columns, 1, "array-only composite must install with its real attribute")
-	require.NotNil(t, cat.GetRelation("public", "users"), "table using a composite column must install")
-}
-
 func TestWalkThroughCompositeFallbackPreservesAttributeNames(t *testing.T) {
 	// A domain-typed attribute cannot install (domains are not loader
 	// objects), forcing the pseudo fallback — which must keep attribute
@@ -308,7 +160,7 @@ func TestWalkThroughCompositeFallbackPreservesAttributeNames(t *testing.T) {
 	}
 
 	cat := catalog.New()
-	err := loadWalkThroughCatalog(context.Background(), cat, meta)
+	_, err := cat.LoadMetadata(context.Background(), meta, catalog.LoadMetadataOptions{Full: true})
 	require.NoError(t, err)
 
 	rel := cat.GetRelation("public", "with_domain")
@@ -373,7 +225,7 @@ func TestWalkThroughDropReaddAttributeReadsCatalogType(t *testing.T) {
 	}
 
 	cat := catalog.New()
-	err := loadWalkThroughCatalog(context.Background(), cat, meta)
+	_, err := cat.LoadMetadata(context.Background(), meta, catalog.LoadMetadataOptions{Full: true})
 	require.NoError(t, err)
 
 	catAfter := cat.Clone()
@@ -412,7 +264,7 @@ func TestWalkThroughAppliesCompositeTypeChanges(t *testing.T) {
 	meta := compositeWalkThroughMetadata()
 
 	catBefore := catalog.New()
-	err := loadWalkThroughCatalog(context.Background(), catBefore, meta)
+	_, err := catBefore.LoadMetadata(context.Background(), meta, catalog.LoadMetadataOptions{Full: true})
 	require.NoError(t, err)
 
 	catAfter := catBefore.Clone()
