+++ backend/plugin/schema/pg/walk_through_test.go
@@ -1,12 +1,14 @@
 package pg
 
 import (
+	"context"
 	"io"
 	"os"
 	"path/filepath"
 	"testing"
 
 	metadatapb "github.com/bytebase/omni/metadata"
+	"github.com/bytebase/omni/pg/catalog"
 	"github.com/google/go-cmp/cmp"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/proto"
@@ -368,4 +370,87 @@ func newSearchPathTestTable(name string) *metadatapb.TableMetadata {
 			},
 		},
 	}
+}
+
+// TestClone_SearchPath tests Clone preserves search path and session user.
+func TestClone_SearchPath(t *testing.T) {
+	meta := &metadatapb.DatabaseSchemaMetadata{
+		Name: "postgres",
+		Schemas: []*metadatapb.SchemaMetadata{
+			{Name: "public"},
+			{Name: "alice"},
+		},
+	}
+
+	catBefore := catalog.New()
+	catBefore.SetSessionUser("alice")
+	catBefore.SetSearchPath([]string{"$user", "public"})
+	_, err := catBefore.LoadMetadata(context.Background(), meta, catalog.LoadMetadataOptions{Full: true})
+	require.NoError(t, err)
+
+	catAfter := catBefore.Clone()
+
+	// CREATE TABLE without schema should use search path ($user → alice)
+	_, err = catAfter.Exec(`CREATE TABLE my_table (id int);`, nil)
+	require.NoError(t, err)
+
+	// Table should be in alice schema on clone
+	require.NotNil(t, catAfter.GetRelation("alice", "my_table"))
+
+	// Original should NOT have it
+	require.Nil(t, catBefore.GetRelation("alice", "my_table"))
+
+	// Diff should show the new table
+	diff := catalog.Diff(catBefore, catAfter)
+	require.False(t, diff.IsEmpty())
+
+	foundAliceTable := false
+	for _, rel := range diff.Relations {
+		if rel.SchemaName == "alice" && rel.Name == "my_table" && rel.Action == catalog.DiffAdd {
+			foundAliceTable = true
+		}
+	}
+	require.True(t, foundAliceTable, "diff should show alice.my_table as added")
+}
+
+// TestClone_WalkThroughFunction tests the actual WalkThroughWithContext function
+// using Clone (if enabled) produces correct FinalMetadata.
+func TestClone_WalkThroughFunction(t *testing.T) {
+	meta := &metadatapb.DatabaseSchemaMetadata{
+		Name: "postgres",
+		Schemas: []*metadatapb.SchemaMetadata{
+			{
+				Name: "public",
+				Tables: []*metadatapb.TableMetadata{
+					{
+						Name: "test",
+						Columns: []*metadatapb.ColumnMetadata{
+							{Name: "id", Type: "integer", Position: 1},
+							{Name: "name", Type: "text", Position: 2, Nullable: true},
+						},
+						Indexes: []*metadatapb.IndexMetadata{
+							{Name: "test_pkey", Expressions: []string{"id"}, Unique: true, Primary: true},
+						},
+					},
+				},
+			},
+		},
+	}
+
+	state := model.NewDatabaseMetadata(meta, nil, nil, storepb.Engine_POSTGRES, true)
+	ctx := schema.WalkThroughContext{
+		RawSQL: `CREATE TABLE public.new_table (id int PRIMARY KEY, val text);`,
+	}
+
+	advice := WalkThroughWithContext(ctx, state, nil)
+	require.Nil(t, advice, "walk-through should succeed")
+
+	// Check FinalMetadata has the new table
+	newTbl := state.GetSchemaMetadata("public").GetTable("new_table")
+	require.NotNil(t, newTbl, "new_table should exist in FinalMetadata")
+
+	// Check original table is preserved
+	origTbl := state.GetSchemaMetadata("public").GetTable("test")
+	require.NotNil(t, origTbl, "test table should still exist")
+	require.Equal(t, 2, len(origTbl.GetProto().Columns))
 }
